+++ common/jinja2/common/preview-workbasket.jinja
@@ -32,6 +32,6 @@
       </ul>
     </div>
   </div>
-  <a href="{{ url('review-workbasket') }}" class="govuk-button govuk-button--primary">Review workbasket</a>
-  <a href="{{ url('edit-workbasket') }}?edit=1" class="govuk-button govuk-button--primary">Edit workbasket</a>
+  <a href="{{ url('workbaskets:review-workbasket') }}" class="govuk-button govuk-button--primary">Review workbasket</a>
+  <a href="{{ url('workbaskets:edit-workbasket') }}?edit=1" class="govuk-button govuk-button--primary">Edit workbasket</a>
 {% endblock %}
