+++ common/urls.py
@@ -15,21 +15,6 @@ register_converter(NumericSIDConverter, "sid")
 
 urlpatterns = [
     path("dashboard", views.DashboardView.as_view(), name="dashboard"),
-    path(
-        "preview-workbasket/",
-        views.PreviewWorkbasketView.as_view(),
-        name="preview-workbasket",
-    ),
-    path(
-        "edit-workbasket/",
-        views.EditWorkbasketView.as_view(),
-        name="edit-workbasket",
-    ),
-    path(
-        "review-workbasket/",
-        views.ReviewWorkbasketView.as_view(),
-        name="review-workbasket",
-    ),
     path("", views.WorkbasketActionView.as_view(), name="index"),
     path("healthcheck", views.healthcheck, name="healthcheck"),
     path("login", views.LoginView.as_view(), name="login"),
