+++ workbaskets/urls.py
@@ -22,6 +22,26 @@ ui_patterns = [
         name="workbasket-ui-create",
     ),
     path(
+        "preview-workbasket/",
+        ui_views.PreviewWorkbasketView.as_view(),
+        name="preview-workbasket",
+    ),
+    path(
+        "edit-workbasket/",
+        ui_views.EditWorkbasketView.as_view(),
+        name="edit-workbasket",
+    ),
+    path(
+        "review-workbasket/",
+        ui_views.ReviewWorkbasketView.as_view(),
+        name="review-workbasket",
+    ),
+    path(
+        "download",
+        ui_views.download_envelope,
+        name="workbasket-download",
+    ),
+    path(
         f"<pk>/",
         ui_views.WorkBasketDetail.as_view(),
         name="workbasket-ui-detail",
@@ -45,11 +65,6 @@ ui_patterns = [
         f"<pk>/delete-changes-done/",
         ui_views.WorkBasketDeleteChangesDone.as_view(),
         name="workbasket-ui-delete-changes-done",
-    ),
-    path(
-        "download",
-        ui_views.download_envelope,
-        name="workbasket-download",
     ),
 ]
